main activity; courses and profile fragment
