+++ android/app/src/main/java/mannschaft_knust/classrep/MainActivity.java
@@ -1,95 +1,129 @@
 package mannschaft_knust.classrep;
 
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.support.transition.AutoTransition;
-import android.support.transition.Scene;
-import android.support.transition.Transition;
-import android.support.transition.TransitionManager;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.text.InputType;
+import android.support.annotation.NonNull;
+import android.support.design.widget.BottomNavigationView;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v4.app.ListFragment;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
 
-    private ViewGroup sceneRoot;
-    private Scene scene1;
-    private Scene scene2;
-    private Transition autoTransition;
-    private View clickedView;
-    private Scene currentScene;
+    ProfileFragment profileFragment = new ProfileFragment();
+    FragmentManager fragmentManager = getSupportFragmentManager();
+
+    private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener
+            = new BottomNavigationView.OnNavigationItemSelectedListener() {
+
+        @Override
+        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+            FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+            switch (item.getItemId()) {
+                case R.id.navigation_courses:
+                    fragmentTransaction.remove(profileFragment);
+                    fragmentTransaction.commit();
+                    return true;
+                case R.id.navigation_user_profile:
+                    fragmentTransaction.replace(R.id.profile_fragment_container, profileFragment);
+                    fragmentTransaction.commit();
+                    return true;
+            }
+            return false;
+        }
+    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
 
-        SharedPreferences userPref;
+        /*SharedPreferences userPref;
         userPref = getSharedPreferences(
                 "mannschaft_knust.classrep.USER_PREF" , MODE_PRIVATE);
-        if (userPref.contains("username") && userPref.contains("password"))
-            //code for loading courses activity here
-            ;
-        else {
-            super.onCreate(savedInstanceState);
-            setContentView(R.layout.activity_main);
+        if ( !(userPref.contains("username") && userPref.contains("password"))){
+            Intent signInIntent = new Intent(this, SignInActivity.class);
+            startActivity(signInIntent);
+        }*/
 
-            //scenes and transition
-            sceneRoot = findViewById(R.id.scene_root);
-            scene1 = Scene.getSceneForLayout(sceneRoot, R.layout.activity_main_scene1,
-                    this);
-            scene2 = Scene.getSceneForLayout(sceneRoot, R.layout.activity_main_scene2,
-                    this);
-            scene2.setEnterAction(new Runnable() {
-                @Override
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
 
-                //do this every time whe show sign in view
-                public void run() {
-                    EditText userIDInput = findViewById(R.id.userID);
-                    TextView signUpLink = findViewById(R.id.signUpLink);
+        Toolbar actionBar = findViewById(R.id.action_bar);
+        setSupportActionBar(actionBar);
 
-                    if(clickedView.getId() == R.id.student_button){
-                        userIDInput.setHint(R.string.student_id);
-                        userIDInput.setInputType(InputType.TYPE_CLASS_NUMBER);
-                        signUpLink.setClickable(true);
-                        signUpLink.setVisibility(View.VISIBLE);
-                    }
-                    else if(clickedView.getId() == R.id.instructor_button){
-                        userIDInput.setHint(R.string.instructor_id);
-                        userIDInput.setInputType(InputType.TYPE_TEXT_VARIATION_EMAIL_ADDRESS);
-                        signUpLink.setClickable(false);
-                        signUpLink.setVisibility(View.INVISIBLE);
-                    }
-                }
-            });
-            autoTransition = new AutoTransition();
-        }
-    }
+        BottomNavigationView navigation = findViewById(R.id.navigation);
+        navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
 
-    // called when user typ(instructor or student button is clicked
-    public void onClickUserType(View button){
-        clickedView = button;
+        CoursesListFragment coursesListFragment = new CoursesListFragment();
+        CourseViewFragment courseViewFragment = new CourseViewFragment();
 
-        //transition to sign in view
-        TransitionManager.go(scene2, autoTransition);
-        currentScene = scene2;
     }
 
     @Override
-    public void onBackPressed() {
-        //go back to scene1 if back pressed on scene 2
-        if (currentScene == scene2) {
-            TransitionManager.go(scene1, autoTransition);
-            currentScene = scene1;
-        }
-        else
-            super.onBackPressed();
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_action_bar, menu);
+
+        //MenuItem searchItem = menu.findItem(R.id.action_search);
+        //SearchView searchView = (SearchView) searchItem.getActionView();
+
+        // Configure the search info and add any event listeners...
+
+        return super.onCreateOptionsMenu(menu);
     }
 
-    public void onClickSignUpLink(View signUpLink){
-        Intent signUpIntent = new Intent(this, SignUpActivity.class);
-        startActivity(signUpIntent);
+    /**
+     * Created by Hassan on 22/03/2018.
+     */
+
+    public static class CoursesListFragment extends ListFragment {
+
+        @Override
+        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState){
+            // Inflate the layout for this fragment
+            return inflater.inflate(R.layout.courses_list_fragment, container, false);
+        }
     }
+
+    /**
+     * Created by Hassan on 23/03/2018.
+     */
+
+    public static class CourseViewFragment extends ListFragment {
+        @Override
+        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState){
+            // Inflate the layout for this fragment
+            return inflater.inflate(R.layout.course_view_fragment, container, false);
+        }
+    }
+
+    /**
+     * Created by Hassan on 23/03/2018.
+     */
+
+    public static class PostViewFragment extends Fragment {
+        @Override
+        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState){
+            // Inflate the layout for this fragment
+            return inflater.inflate(R.layout.post_view_fragment, container, false);
+        }
+    }
+
+    /**
+     * Created by Hassan on 22/03/2018.
+     */
+
+    public static class ProfileFragment extends Fragment {
+        @Override
+        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState){
+            // Inflate the layout for this fragment
+            return inflater.inflate(R.layout.profile_fragment, container, false);
+        }
+    }
+
 }
