+++ android/app/src/main/res/color/color_state_list.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true"
+        android:color="@color/colorPrimary" />
+    <item android:color="#FF000000" />
+</selector>
