+++ android/app/src/main/res/menu/navigation_instructor.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_courses"
+        android:icon="@drawable/ic_assignment_black_24dp"
+        android:title="@string/courses" />
+
+    <item
+        android:id="@+id/navigation_user_profile"
+        android:icon="@drawable/ic_person_outline_black_24dp"
+        android:title="@string/profile" />
+
+</menu>
