+++ android/app/src/main/res/values/strings.xml
@@ -1,9 +1,8 @@
 <resources>
     <string name="app_name">ClassRep</string>
-    <string name="user_student">Student</string>
-    <string name="user_instructor">Instructor</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="student">Student</string>
+    <string name="instructor">Instructor</string>
+    <string name="profile">Profile</string>
     <string name="or">or</string>
     <string name="sign_in">Sign In</string>
     <string name="sign_up">Sign Up</string>
@@ -17,5 +16,7 @@
     <string name="year">Year</string>
     <string name="retype_password">retype password</string>
     <string name="title_activity_courses">CoursesActivity</string>
-    <string name="title_home">Home</string>
+    <string name="courses">Courses</string>
+    <string name="timetable">Timetable</string>
+    <string name="search">Search</string>
 </resources>
